Add files via upload
 NMS-Cinemas 
This document contains sections for: 
Description of the project 
Core concepts used in project
Flow of the Application.
Demonstrating the product capabilities, appearance, and user interactions.


			Description of the project
Create a dynamic and responsive web application for booking movie tickets online for different genres and languages.
Background of the problem statement: NMS Cinemas is a chain of single screen theatres that screen movie shows of different genres and languages at very genuine prices. It was established in 2004 in Pune, India. Recently, the business analysts noticed a decline in sales since 2010. They found out that the online booking of movie tickets from apps, such as BookMyShow and Paytm were gaining more profit by eliminating middlemen from the equation. As a result, the team decided to hire a Full Stack developer to develop an online movie ticket booking web application with a rich and user-friendly interface. You are hired as the Full Stack Java developer and are asked to develop the web application. The management team has provided you with the requirements and their business model so that you can easily arrange different components of the application.






Core concepts used in the Project

Database management: MySQL and Oracle
Backend logic: Java programming, NodeJS
Frontend development: JSP, Angular, Bootstrap, HTML/CSS, and Javascript
Automation and testing technologies: Selenium, Jasmine, and TestNG
DevOps and production technologies: Git, GitHub, Jenkins, Docker, Kubernetes, and AWS

Sprints planning and Task completion
The project is planned to be completed in 3 sprint. Tasks assumed to be completed in the sprint are:
Creating the flow of the application
Initialising git repository to track changes as development progresses.
Writing the html pages and javascript program to fulfil the requirements of the project.
Testing the pageswith different kinds of User input
Pushing code to GitHub.
Creating this specification document highlighting application capabilities, appearance, and user interactions

All code for this project can be found at the following link: 

https://github.com/sruthisree001/NMS-cinemas.git



Home page

Another section


Loginpage



Ticket booking :selecting the movie name
Seat selection


Price based on number of seats

Payment options



Successful payment page